+++ lib/Model/Record.php
@@ -24,16 +24,6 @@ abstract class Record
     private $type;
 
     /**
-     * @var array<string>
-     */
-    private $implementations = [];
-
-    /**
-     * @var array<string>
-     */
-    private $implemented = [];
-
-    /**
      * @var string
      */
     private $filePath;
@@ -75,56 +65,10 @@ abstract class Record
         return $this;
     }
 
-    public function addImplementation($fqn): void
-    {
-        // TODO: Remove
-        if ($fqn instanceof ReflectionClassLike) {
-            $this->implementations[$fqn->name()->full()] = $fqn->name()->full();
-            return;
-        }
-        $this->implementations[(string)$fqn] = (string)$fqn;
-    }
-
-    public function addImplements($fqn): void
-    {
-        // TODO: Remove
-        if ($fqn instanceof ReflectionClassLike) {
-            $this->implemented[$fqn->name()->full()] = $fqn->name()->full();
-            return;
-        }
-        $this->implemented[(string)$fqn] = (string)$fqn;
-    }
-
-    public function removeClass(FullyQualifiedName $implementedClass): void
-    {
-        foreach ($this->implementations as $key => $implementation) {
-            if ($implementation !== $implementedClass->__toString()) {
-                continue;
-            }
-
-            unset($this->implementations[$key]);
-        }
-    }
 
     public function fqn(): FullyQualifiedName
     {
         return $this->fqn;
-    }
-
-    /**
-     * @return array<string>
-     */
-    public function implementations(): array
-    {
-        return $this->implementations;
-    }
-
-    /**
-     * @return array<string>
-     */
-    public function implementedClasses(): array
-    {
-        return $this->implemented;
     }
 
     public function lastModified(): int
